+++ PythonAPI/carla/source/carla/command.py
@@ -1,8 +1,0 @@
-# Copyright (c) 2019 Computer Vision Center (CVC) at the Universitat Autonoma de
-# Barcelona (UAB).
-#
-# This work is licensed under the terms of the MIT license.
-# For a copy, see <https://opensource.org/licenses/MIT>.
-
-# pylint: disable=W0401
-from .libcarla.command import *